feat(viewer)!: make S3 support optional
Gate the viewer's AWS SDK-backed storage, bring-your-own credentials,
and assume-role support behind a default-on `s3` feature.

Allow `dial9-viewer --no-default-features` to build a local-only viewer
without AWS dependencies. Keep `dial9 --features cli` S3-capable by
explicitly forwarding `dial9-viewer/s3`.

BREAKING CHANGE: consumers that disable `dial9-viewer` default features
must enable `s3` to retain its S3 APIs.
