+++ dial9-viewer/Cargo.toml
@@ -30,7 +30,10 @@ include = [
 pkg-url = "{ repo }/releases/download/{ name }-v{ version }/{ name }-{ target }-v{ version }{ archive-suffix }"
 
 [features]
-dev-server = ["dep:s3s", "dep:s3s-fs", "dep:s3s-aws", "dep:tempfile"]
+default = ["s3"]
+## S3 storage, bring-your-own AWS credentials, and assume-role support.
+s3 = ["dep:aws-config", "dep:aws-sdk-s3"]
+dev-server = ["s3", "dep:s3s", "dep:s3s-fs", "dep:s3s-aws", "dep:tempfile"]
 
 [[bin]]
 name = "dial9-viewer"
@@ -48,6 +51,14 @@ required-features = ["dev-server"]
 name = "gen-fixtures"
 path = "src/bin/gen_fixtures.rs"
 required-features = ["dev-server"]
+
+[[test]]
+name = "aggregate_test"
+required-features = ["s3"]
+
+[[test]]
+name = "server_test"
+required-features = ["s3"]
 
 [dependencies]
 axum = "0.8"
@@ -71,8 +82,8 @@ tracing-subscriber = { version = "0.3", features = ["env-filter", "json"] }
 metrique = { version = "0.1.28", features = ["emf", "local-format"] }
 metrique-writer = "0.1"
 time = "0.3"
-aws-config = { version = "1", features = ["behavior-version-latest"] }
-aws-sdk-s3 = { version = "1", default-features = false, features = ["behavior-version-latest", "default-https-client", "http-1x", "rt-tokio", "sigv4a"] }
+aws-config = { version = "1", features = ["behavior-version-latest"], optional = true }
+aws-sdk-s3 = { version = "1", default-features = false, features = ["behavior-version-latest", "default-https-client", "http-1x", "rt-tokio", "sigv4a"], optional = true }
 anyhow = "1"
 bon = "3"
 s3s = { version = "0.14.1", optional = true }
